refactor: extract selfTest util to core-test-kit
