+++ packages/core-test-kit/src/index.ts
@@ -4,3 +4,4 @@ export * from './matchers/flat-match';
 export * from './matchers/match-css';
 export * from './matchers/results';
 export * from './match-rules';
+export * from './inline-expectation';
