+++ packages/core/test/stylable-transformer/scope-selector-v2.spec.ts
@@ -1,39 +1,4 @@
-import type * as postcss from 'postcss';
-import { generateStylableRoot } from '@stylable/core-test-kit';
-
-function selfTest(
-    result: postcss.Root,
-    expectedTestsCount = result.toString().match(/@check/gm)!.length
-) {
-    if (expectedTestsCount === 0) {
-        throw new Error('no tests found try to add @check comments before any selector');
-    }
-    const checks: Array<[string, string]> = [];
-
-    result.walkRules((rule) => {
-        const p = rule.prev();
-        if (p && p.type === 'comment') {
-            const m = p.text.match(/@check\s+(.*)/);
-            if (m) {
-                checks.push([rule.selector, m[1]]);
-            }
-        }
-    });
-    const errors: string[] = [];
-    checks.forEach(([a, b]) => {
-        if (a !== b) {
-            errors.push(`expected ${a} to transform to ${b}`);
-        }
-    });
-    if (errors.length) {
-        throw new Error(errors.join('\n'));
-    }
-    if (expectedTestsCount !== checks.length) {
-        throw new Error(
-            `Expected ${expectedTestsCount} checks to run but there was ${checks.length}`
-        );
-    }
-}
+import { generateStylableRoot, testInlineExpects } from '@stylable/core-test-kit';
 
 describe('Stylable scope-selector v2', () => {
     it('should handle basic transform', () => {
@@ -160,7 +125,7 @@ describe('Stylable scope-selector v2', () => {
             },
         });
 
-        selfTest(result);
+        testInlineExpects(result);
     });
 
     it('should properly scope states in nested-pseudo-classes', () => {
@@ -219,7 +184,7 @@ describe('Stylable scope-selector v2', () => {
             },
         });
 
-        selfTest(result);
+        testInlineExpects(result);
     });
 
     it('should properly scope states in nested-pseudo-classes (more exmaples)', () => {
@@ -304,7 +269,7 @@ describe('Stylable scope-selector v2', () => {
             },
         });
 
-        selfTest(result);
+        testInlineExpects(result);
     });
 
     it('should properly scope states in nested-pseudo-classes with aliase state override', () => {
@@ -368,7 +333,7 @@ describe('Stylable scope-selector v2', () => {
             },
         });
 
-        selfTest(result);
+        testInlineExpects(result);
     });
 
     it('should properly scope states in nested-pseudo-classes222231241241242', () => {
@@ -401,6 +366,6 @@ describe('Stylable scope-selector v2', () => {
             },
         });
 
-        selfTest(result);
+        testInlineExpects(result);
     });
 });
